+++ app/src/main/java/org/oxycblt/auxio/playback/state/PlaybackStateManager.kt
@@ -34,6 +34,10 @@ import org.oxycblt.auxio.util.logW
 /**
  * Master class (and possible god object) for the playback state.
  *
+ * Whereas other apps centralize the playback state around the MediaSession, Auxio does not, as
+ * MediaSession is a terrible API that prevents nice features like better album cover loading and a
+ * reasonable queue system.
+ *
  * This should ***NOT*** be used outside of the playback module.
  * - If you want to use the playback state in the UI, use
  * [org.oxycblt.auxio.playback.PlaybackViewModel] as it can withstand volatile UIs.